Move package namespace (#50)
