--- sonar-perl-plugin/src/main/java/com/github/otrosien/sonar/perl/critic/PerlCriticProperties.java
@@ -1,9 +1,0 @@
-package com.github.otrosien.sonar.perl.critic;
-
-public class PerlCriticProperties { // NOSONAR - yes, this is a util class.
-
-    public static final String PERLCRITIC_REPORT_PATH_KEY = "sonar.perlcritic.reportPath";
-
-    public static final String PERLCRITIC_REPORT_PATH_DEFAULT = "perlcritic_report.txt";
-
-}
