--- sonar-perl-plugin/src/main/java/com/github/otrosien/sonar/perl/tap/TestHarnessArchiveProperties.java
@@ -1,9 +1,0 @@
-package com.github.otrosien.sonar.perl.tap;
-
-public class TestHarnessArchiveProperties { // NOSONAR
-
-    public static final String HARNESS_ARCHIVE_PATH_KEY = "sonar.perl.testHarness.archivePath";
-
-    public static final String HARNESS_ARCHIVE_PATH_DEFAULT = "testReport.tgz";
-
-}
