--- sonar-perl-plugin/src/main/java/com/github/otrosien/sonar/perl/tap/TestHarnessJUnitProperties.java
@@ -1,7 +1,0 @@
-package com.github.otrosien.sonar.perl.tap;
-
-public class TestHarnessJUnitProperties { // NOSONAR
-
-    public static final String HARNESS_JUNIT_PATH_KEY = "sonar.perl.testHarness.junitPath";
-
-}
