+++ sonar-perl-plugin/src/main/java/com/github/sonarperl/critic/PerlCriticProperties.java
@@ -1,0 +1,9 @@
+package com.github.sonarperl.critic;
+
+public class PerlCriticProperties { // NOSONAR - yes, this is a util class.
+
+    public static final String PERLCRITIC_REPORT_PATH_KEY = "sonar.perlcritic.reportPath";
+
+    public static final String PERLCRITIC_REPORT_PATH_DEFAULT = "perlcritic_report.txt";
+
+}
