+++ sonar-perl-plugin/src/main/java/com/github/sonarperl/tap/TestHarnessArchiveProperties.java
@@ -1,0 +1,9 @@
+package com.github.sonarperl.tap;
+
+public class TestHarnessArchiveProperties { // NOSONAR
+
+    public static final String HARNESS_ARCHIVE_PATH_KEY = "sonar.perl.testHarness.archivePath";
+
+    public static final String HARNESS_ARCHIVE_PATH_DEFAULT = "testReport.tgz";
+
+}
