+++ sonar-perl-plugin/src/main/java/com/github/sonarperl/tap/TestHarnessJUnitProperties.java
@@ -1,0 +1,7 @@
+package com.github.sonarperl.tap;
+
+public class TestHarnessJUnitProperties { // NOSONAR
+
+    public static final String HARNESS_JUNIT_PATH_KEY = "sonar.perl.testHarness.junitPath";
+
+}
